requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Configuration
pyyaml>=6.0

# HTTP & API
requests>=2.31.0
openai>=1.0.0
python-multipart>=0.0.6
